requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
motor>=3.3.0
pymongo>=4.6.0

# Authentication & security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
cryptography>=41.0.0

# MFA
pyotp>=2.9.0

# HTTP clients
httpx>=0.25.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# YAML parsing
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
